+++ tests/ElementFinderTest.php
@@ -8,7 +8,6 @@ use InvalidArgumentException;
 use RuntimeException;
 use PHPUnit\Framework\TestCase;
 use Test\Xparse\ElementFinder\Dummy\ItemsByClassExpressionTranslator;
-use Xparse\ElementFinder\Collection\StringCollection;
 use Xparse\ElementFinder\DomNodeListAction\RemoveNodes;
 use Xparse\ElementFinder\ElementFinder;
 use Xparse\ElementFinder\ElementFinderInterface;