+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 All Notable changes to `ElementFinder` will be documented in this file
 
 ## Planned changes
+- Add `ExpressionTranslator` (merged from the [xparse/expression-translator](https://github.com/xparse/ExpressionTranslator) repository)
+- Move to php 8.1
 
 ## 1.0.1 [2021-06-01]
-Move xparse/expression-translator to 1.0.0
+- Move xparse/expression-translator to 1.0.0
 
 ## 1.0.0 [2021-05-31]
 - No breaking changes.